Move node selection to gridmodel
